Bump golang 1.21 (#121)
* bump go to 1.21 in Dockerfiles

* bump go to 1.21 in makefile and ci.go

* common: implement old version of func sort.Slice

* replace package sort with common/sort

* not compare strings in func CompareSignersLists

* bump go to 1.21 in go.mod & go.sum

* console, internal/jsre: use github.com/dop251/goja (#20470)

This replaces the JavaScript interpreter used by the console with goja,
which is actively maintained and a lot faster than otto. Clef still uses otto
and eth/tracers still uses duktape, so we are currently dependent on three
different JS interpreters. We're looking to replace the remaining uses of otto
soon though.

* Upgrade go.mod according to XDPoSChain

* Bump golang 1.21 in github workflow

* disable buildvcs (default changed to true in new go version)

---------

Co-authored-by: Daniel Liu <liudaniel@qq.com>
Co-authored-by: Guillaume Ballet <gballet@gmail.com>
Co-authored-by: wanwiset25 <wanwiset25@gmail.com>